Inherit font-family in input, textarea, and button
In browsers like Chrome and Firefox, font-families of `input`, `textarea`, and `button` are
initially set by the browsers.  For example, even if you set `font-family` for `body`, fonts of
`input`, `textarea`, and `button` may be different.  While the different-fonts itself is not a
problem, sometimes it introduces mis-aligned input boxes and buttons.

As a workaround, this commit forces to inherit font-family in the elements.